:recycle: api: Move EmailingAttempt to src
